Delete suffix 'Entity' + formatting
