+++ backend/src/main/java/ru/hh/school/checkupextension/core/data/entity/SubmissionShortInfo.java
@@ -1,0 +1,3 @@
+package ru.hh.school.checkupextension.core.data.entity;
+
+public record SubmissionShortInfo(long submissionId, byte status) { }
