--- backend/src/main/java/ru/hh/school/checkupextension/core/data/entity/SubmissionShortInfoEntity.java
@@ -1,3 +1,0 @@
-package ru.hh.school.checkupextension.core.data.entity;
-
-public record SubmissionShortInfoEntity(long submissionId, byte status) { }
