+++ backend/src/main/java/ru/hh/school/checkupextension/core/data/entity/Verification.java
@@ -1,0 +1,75 @@
+package ru.hh.school.checkupextension.core.data.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+
+import java.util.Objects;
+
+@Entity
+@Table(name = "verification")
+public class Verification {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "id")
+  private Long id;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "problem_id")
+  private Problem problem;
+
+  @Column(name = "content")
+  private String content;
+
+  public Verification() {
+  }
+
+  public Long getId() {
+    return id;
+  }
+
+  public void setId(Long id) {
+    this.id = id;
+  }
+
+  public Problem getProblem() {
+    return problem;
+  }
+
+  public void setProblem(Problem problem) {
+    this.problem = problem;
+  }
+
+  public String getContent() {
+    return content;
+  }
+
+  public void setContent(String content) {
+    this.content = content;
+  }
+
+  @Override
+  public boolean equals(Object o) {
+      if (this == o) {
+          return true;
+      }
+      if (o == null || getClass() != o.getClass()) {
+          return false;
+      }
+    Verification that = (Verification) o;
+    return Objects.equals(id, that.id) && Objects.equals(problem, that.problem)
+        && Objects.equals(content, that.content);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(id, problem, content);
+  }
+}
