--- backend/src/main/java/ru/hh/school/checkupextension/core/data/entity/VerificationEntity.java
@@ -1,71 +1,0 @@
-package ru.hh.school.checkupextension.core.data.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-
-import java.util.Objects;
-
-@Entity
-@Table(name = "verification")
-public class VerificationEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "problem_id")
-    private ProblemEntity problem;
-
-    @Column(name = "content")
-    private String content;
-
-    public VerificationEntity() {
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public ProblemEntity getProblem() {
-        return problem;
-    }
-
-    public void setProblem(ProblemEntity problem) {
-        this.problem = problem;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        VerificationEntity that = (VerificationEntity) o;
-        return Objects.equals(id, that.id) && Objects.equals(problem, that.problem)
-                && Objects.equals(content, that.content);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, problem, content);
-    }
-}
